Adaptive W-TinyLfu [skip ci]
Thanks for users, we have more extreme cases where TinyLfu performs
poorly due to recency-biased workloads. This motivated the admission
window to capture short bursts. However this was set to a generally
good value of 1% that worked well in the majority of traces. Originally
we only saw a larger value (20%) modestly improved some cases, like
ARC's and UMass' OLTP traces. The Scarab, Gradle, and Corda traces show
much more bias and require larger windows. Yet doing so statically would
hurt frequency-biased traces like db, search, and analytics.

We use hill climbing to optimize the admission window size. This is done
by sampling the hit rate, adjusting the setting, and repeating in the
same direction if improved or turning around. We initially take large
steps to combat noise and decay it so that over many iterations we
converge (rather than occilating around the peak). This process is
restarted whenever the hit rate changes by >= 5@, indicating that the
workload may have changed. Since some users have very large caches, we
amortize the transfer of items over multiple cycles to avoid penalizing
a thread.

A minor change of note is that many previously lazySet volatile fields
are now not volatile. These were almost exclusively read or written to
under the eviction lock, so it was an oversight when writing the code
generator.

This commit appears larger due to surrounding cleanup. In particular the
very early name "eden" was replaced with "window", the name we chose but
hadn't updated to. The comment separator was also made balanced, and
amusingly was copied from ConcurrentHashMap where the typos remain.